requirements.txt
httpx
pytz
python-dotenv